Merge in tuning changes to the Kp value for better consistency.

+++ pick_and_place_manipulator/psoc_creator/robogrok1/pick_and_place_manipulator.cydsn/servos.h
@@ -7,8 +7,8 @@
     
 #include "project.h"
     
-float theta_1(float angle);
-float theta_2(float angle);
+int theta_1(float angle);
+int theta_2(float angle);
 void servo_1_position_degrees(float angle, int delay);
 void servo_2_position_degrees(float angle, int delay);
 void servos_position_reset(int delay);
